requirements.txt
boto3
idna
Flask
flask-accepts
Flask-Cors
flask-marshmallow
Flask==2.2.2
flask-accepts==0.18.4
Flask-Cors==3.0.10
flask-marshmallow==0.14.0
Flask-PyMongo==2.3.0
flask-restx
gunicorn
pymongo
python-dateutil
PyJWT
Werkzeug==2.2.2
marshmallow-sqlalchemy
xxhash
